+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for PCML generation
 - removed restriction on Address.cs nullflavor
 - Added Hl7v2 DI to CDA functionality
+- Fixed HL7 requester.ParticipationEndTime to OBR-27.5 rather than OBR-22
 
 4.4.2 - Reference Platform Clinical Document Library (Feb 2018)
 ----------------------------------------------------